Keep correct txNum

+++ cmd/state/commands/erigon2.go
@@ -140,7 +140,7 @@ func Erigon2(genesis *core.Genesis, logger log.Logger) error {
 			}
 		}
 	}
-	blockNum := 0
+	blockNum := uint64(0)
 	var txNum uint64 = 1
 	trace := false
 	logEvery := time.NewTicker(logInterval)
